reyk@: www/mozilla-firefox U2F/FIDO support - using fido(4)
https://marc.info/?l=openbsd-ports&m=157666694220206&q=mbox

On Tue, Dec 10, 2019 at 01:57:42PM +0000, Reyk Floeter wrote:
> TL;DR: this is an experimental port update to add U2F/FIDO support and
=2E..
> I've implemented the missing bits in the Rust crate that is
> responsible for accessing the USB HID devices, see
> https://github.com/reyk/authenticator-rs/tree/reyk/openbsd
>=20

This updated version has some changes:
1. It uses the new fido(4) driver in OpenBSD-current instead of uhid(4).
2. It works under pledge(2), so it doesn't need any changes in pledge.main.
3. It works better under unveil(2) with the /dev/fido/ directory.
4. It fixes a TOCTOU problem that was given by the authenticator-rs design:
   the existing OS backends (such as freebsd) first scan for devices
   (in monitor) before they pass the device name (e.g. /dev/uhid0) to
   the Device struct that opens it on demand.  I've had the same approach
   in my earlier diff but now I open the devices in monitor and pass the
   fd to the Device struct instead of the (totally pointless) device name.

To test this diff, you need to upgrade to -current or to the latest
snapshot (Dec 18 or newer).

Known issues:

- If you're in the middle of the authentication (Firefox shows a
little fingerprint in the status bar), you cannot use SSH with U2F at
the same time.  The device can only be opened by one user at a time
and sometimes blocks it for a bit too long.  But once you've done the
authentication, firefox closes the device again and you should be fine
and there's no need to kill firefox for SSH.
- The device sometimes doesn't respond immediately.
- I didn't manage to run firefox with RUST_LOG=3Ddebug to see the
debug!() output of the authenticator module.  Any hints would be
welcome.

I'm working on an additional approach on top of fido(4) that
eventually helps to fix or mitigate the known issues.  So this diff
might not be the final concept or version, and I will not try to
upstream it yet.

Please test, for example:
https://webauthn.io/, first Register, then Login.

Reyk